[Update] 사용자의 구매내역 및 판매내역 조회 api 분리 (roki 코드리뷰 반영)
